requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.0.0
simpleeval>=0.9.0

